// src/app/shared/model/projectCreationModel.ts
export interface ProjectCreation {
    id: Number;
    requestNo: string;
    requestLetterNo: string;
    requestLetterDate: string;
    projectCode: string;
    projectName: string;
    projectYear: string;
    agencyId: Number;
    agencyName:string;
    efileNo: string;
    prizmId: string;
    status: string;
    remarks: string;
    nicsiprojectCode: string;
    nicsipino: string;
    pidate: string;
    piamount: Number;
    submitTime: string;
    ipaddress: string;
    submitBy: string;
    modifyBy: string;
    modifyOn: string;
    isActive: string;
    isWorkOrderRequired:boolean;
    isRequestAvailable:string;
}